Refactoring validation functions
